+++ tests/Tools/SinusGeneratorFilter.cpp
@@ -45,19 +45,19 @@ BOOST_AUTO_TEST_CASE( SinusGeneratorFilter_frequency_range_test )
 
 BOOST_AUTO_TEST_CASE( SinusGeneratorFilter_1k_test )
 {
-  ATK::SimpleSinusGeneratorFilter<float> generator;
+  ATK::SimpleSinusGeneratorFilter<double> generator;
   generator.set_output_sampling_rate(48000);
   generator.set_amplitude(-1);
   generator.set_frequency(1000);
   
-  ATK::SinusGeneratorFilter<float> filter;
+  ATK::SinusGeneratorFilter<double> filter;
   filter.set_output_sampling_rate(48000);
   filter.set_frequency(1000);
 
-  ATK::SumFilter<float> sumfilter;
+  ATK::SumFilter<double> sumfilter;
   sumfilter.set_input_sampling_rate(48000);
   
-  ATK::TriangleCheckerFilter<float> checker;
+  ATK::TriangleCheckerFilter<double> checker;
   checker.set_input_sampling_rate(48000);
   checker.set_amplitude(0);
   checker.set_frequency(1000);
